Add files via upload
Write a rotation function which takes as its arguments (i) a vector and (ii) an angle and returns a vector which is the original vector rotated by the given angle. This function should construct and make use of the rotation matrix (for matrix multiplications you can use the intrinsic matmul function).
You should check that your function is working by (i) calling it with three separate points representing a triangle with corners at (1,1), (2,1) and (2,2), (ii) transforming each of these points with a rotation by an angle chosen by the user of the program, and (iii) plotting the initial and final triangles using gnuplot. To plot a single triangle in gnuplot, you should write two columns to a file, with the (x, y) coordinate pairs for the vertices, featuring four rows, with the first point repeated at the end so that the “curve” plotted by gnuplot joins up to form a closed shape. To plot two triangles, you can write four columns, with the (x, y) coordinates of the first triangle in the first two columns, and the second triangles in the third and fourth columns. Your program should write such an output file, calling it rotate.dat. The format of this file will be checked as part of the marking of the assignment. It will be assumed that you have named it as instructed.


Write a translation function which takes as arguments (i) a vector representing an initial point, (ii) an x-displacement and a y-displacement, and (iii) returns a vector describing the translated point.
Note that it may appear simpler to just add the displacements to the coordinates in the vector, but you should generate the translation matrix, as above, and apply it. The reason to do this is that by building a library of (rotation, translation, expansion, reflection, etc.) transformation matrices, you can then build up more complicated transformation operations (e.g. rotate – translate – rotate - reflect ...) by matrix-matrix multiplications, in a way that would not be so straightforward if you did not represent the different individual transformations in the simpler way.
Satisfy yourself that your translation function works by testing it with the triangle used in part 2 and producing an output file translate.dat for gnuplot.

Write a new function which will rotate a triangle about its centre of mass by (i) taking as arguments the three points of a triangle, (ii) translating it so that its centre of mass has moved to the origin, (iii) rotating it through an angle specified by the user of your program, and (iv) translating back, so that its original centre of mass position is restored. For the purposes of completing the assignment, you should take the initial triangle as that defined in section 3.1 (Step 1).
You should write out the coordinates of your original and rotated triangles to a data file and write a gnuplot script (as you did in exercise 8 last semester) that will generate a .png file showing the original and the rotated triangles.