Add files via upload
Your task is to write a program, which will output to a file (called taylor.dat) values of the variable x, the exact function f(x) and with a Taylor approximation up to nth order for a user-specified n value. The output should be in columns so that the result can be plotted using gnuplot. Calculate the function in the range, between x=−2π and x=+2π. As a rough guide, with n=15, your Taylor approximation will look very close to the actual function for the range of x values suggested, but with some noticeable deviations and the ends of the range.
You can structure your program how you choose, though it is suggested that you use a function to calculate the Taylor expansion expression for f(x) to order n, with arguments n and x. If you wish to break the procedure down into more steps, start with a flow chart or a pseudocode outline.