Add files via upload
Linear algebra deals with the manipulation and solution of systems of linear equations. For example, in the case of three simultaneous linear equations in three unknowns, x1 , x2 , and x3 , we may write
a11x1 +a12x2 +a13x3 =b1
a21x1 +a22x2 +a23x3 =b2 (1)
a31x1 +a32x2 +a33x3 =b3
where the coefficients aij and the bk are constants. Such equations can be concisely
written in matrix notation:
€
This matrix A (of numbers aij) where i labels the row and j labels the column, is
expressible in Fortran as an array. In this case it is a 3x3 array since our example
has three columns and three rows. We denote/define such a matrix using the real
variable declaration
real, dimension(1:3,1:3) :: a
which defines an array named a whose element aij is accessed as aij=a(i,j).
Once in matrix notation, the problem of solving the three simultaneous equations in three unknowns becomes that of finding the inverse (A-1) of the matrix A. Applying this inverse to both sides of equation (2) then gives an equation for the three unknowns. It is the purpose of this assignment to write a FORTRAN program to solve such sets of three simultaneous linear equations in three unknowns, making use of matrix algebra.
Step 1: Reading in the matrix
The first step in creating the program is to read in an array from a file. If you find this tricky, re-read the “Higher-dimensional arrays” section of Unit 4 in the course website. Create a file called matrix.dat1 containing the following matrix, laid out in three rows of three:
 €
" a11 a12 a13 %" x1 % " b1 %
$a a a'$x'=$b' (2)
$ 21 22 23'$ 2' $ 2' $a a a'$x'$b' # 31 32 33&# 3& # 3&
#2 1 −1& %(
A=%1 2 1(. %−1 3 2(
$'
Check that you can read in the matrix and write it out again.
1€
Please use this name – it will make compiling and checking the programs much
simpler.
 Continued overleaf...
Step 2: Matrix determinant
We will need to be able to calculate the determinant of a 3x3 array. Write a function which takes as an argument a 3x3 array, calculates the determinant of the associated matrix and returns that single number as the value of the function.
Check that your program is calculating the determinant correctly by calling your function with the array read in during step 1, and print the answer to the screen. Check against the correct value, calculated by hand.
Step 3: Matrix cofactor
Now we will set up a new matrix, which contains the cofactors of the matrix A. The cofactor matrix, C, will be the same size as A (i.e. 3x3) and each element of C is defined such that cij is (-1)i+j multiplied by the determinant of the 2x2 matrix that is obtained by striking out the ith row and jth column for the matrix. For example, given our matrix A:
#2 1 −1& A=%1 2 1( $−1 3 2'
the cofactor c11 of the element a11 is (-1)1+1 = 1, multiplied by the determinant of the
"2 1%
2x2matrix $ ',whichis2x2–3x1=1. Sotheelementc11 is1x1=1.
#32& €
Write a subroutine which takes a matrix as one argument, and returns its cofactor matrix in another argument. From the program, call this subroutine and print the c o€ f a c t o r m a t r i x t o t h e s c r e e n .
Step 4: Matrix Inverse
The inverse of a matrix can be found using
A−1= 1 CT det A
where det A denotes the determinant of matrix A and CT is the transpose of the cofactor matrix C. Use your previously calculated determinant and cofactor to calculate the inverse of m€atrix A. Write this to the screen. Note that Fortran has a transpose intrinsic function.
You can check that the inverse has been calculated correctly by multiplying the inverse by the original matrix. To multiply arrays together according to the usual rules of matrix multiplication, use the matmul intrinsic function.
Step 5: Solving simultaneous equations
Look back at equation (2). If you act on both sides of this equation, from the left, with
-1
A , you get an equation for the unknowns x1, x2, x3. Use this fact to solve the
following system of simultaneous linear equations for x1, x2, and x3:
Last updated 1st of February 2017

2x +x −x =0 123
x1+2x2+x3=3 . −x1 +3x2 +2x3 =1

Print your answer to the screen. When you have finished your program, submit via SurreyLearn, as usual.